+++ Changelog.md
@@ -5,6 +5,7 @@
 * Merged `ParserConfig2` back into `ParserConfig`
 * Added option to the writer to pass through XML markup unmodified
 * `xml-analyze` binary has been moved to examples
+* Writer escapes `--` in comments and `]]>` in CDATA
 
 ## Version 0.8.27